- refactoring resolvers functions.
